Intrinsic fixes; arithmetic intrinsics
Adds all the intrinsic definitions via a macro heavy rust_intrinsics.h
We can add more intrinsic definitions/macros to this file as needed.

It also turns out the template system was broken as fuck. Fixed that and
updated the tests to reflect this change.

Current temp-lifting/variable-renaming pass breaks pattern matching, so
the second intrinsic test won't work, but that's life.

To get the arithmetic intrinsics to work I moved around the location of
things in the generated C: function sigs now come after type definitions.
This is because for intrinsics (macro/template) the "sig" is expanded
immediately into the definition. This is ugly, but it saves us from having
to include the type signatures of the template/macro functions in the
instrinsic defintions.