Makes crust_init optional
If the -optional-init flag is passed to crust, crust no longer uses
crust_init in codegen or code analysis. Instead, all monomorphization
is done using an empty initial set, and the driver function isn't compiled
at all.

This is primarily for us to use in test cases where we want to test some
other feature of the compilation process (currently we always
end up having to write something like crust_init -> (int,) which is
annoying).