Fixes type matching
Turns out the "hacky" approach was indeed too hacky. The problem was requiring the matcher
to produce a value of type T if we were interested in Adt<T>: it's very possible for
Adt<Foo> to be in our universe, but not Foo. So we only fall back on the shadow type variables
if we detect an argument is an abstract type: in that case it seems a good heuristic to
require that all type variables be instantiated to known types

(The above requirement isn't strictly speaking true: you can have only values of Adt<Foo> but
still know about the type Foo. I don't think this will be a problem in practice...)