adds pp to fix crust_init
The code we have right now looks for crust_init, but the name mangling
makes it [file]_crust_init. Rather than change this underlying assumption
I wrote a preprocessor pass to rename it to crust_init

Also, as a temporary measure (until we can figure out the panic macro)
this commit also adds a pass to unmangle crust_abort: it turns out this
is pretty hard-coded into the analysis and code-gen. It's expected that
this pass will be removed after we figure out panic!() and so on.