adds supportr for new pattern forms
The ref match is potentially broken. It will lift the matchee into a
temporary unless the matchee is exactly a variable. I guess this is
to prevent calling a function call several times. I'm not sure what
Rust does. If necessary I can only hoist the matchee to a temporary
if the matchee doesn't have a function call or some other non-side
effecting operation (simple walk of the expr ast).