adds support for vec literals
This was ugly. Turns out that C is pretty picky about where you put
your array literals. In short, they can only be on the rhs of a variable
initialization (the rhs of generic assignments is a no-go). Pointers
and array literals can be interchangeable, but only in one direction.

So, to address this, crust's internal CRep IR added two new block level
statements: vec init and vec assign. Vec init is a special form of let
that admits vec initializers. Vec_assign is a poor mans memmove: it
assigns element-wise from one array to another. The generated C code
is absolutely horrific, but it actually works pretty well.

We don't support returning fixed vecs from functions, because C doesn't
allow that, and if you look at the LLVM IR generated by Rust they
actually add a special "return" pointer for fixed vecs that can't
fit in 64 bits. I have no desire to try to instrument that, so
that's not a feature we'll be supporting in the near future...