adds support for static code gen
Adds code gen for statics. The set of expressions allowed in a static
expression is given by the following grammar:

static ::= var
        | lit
	| & static
	| * static
	| struct { field = static, ... }
	| EVariant(static, static, ...)
	| (static, static, ...)
	| [static, static, ... ]

that is, variables, literals, literal vectors, literal structs/enums,
literal tuples, and address of/deref.

As part of the compilation process crust will compute the necessary static
variables and dump in the definitions as appropriate.