filters test cases that are buggy
The scrubbing script is pretty brittle: it assumes a very specific
format for rustc error messages and further assumes that the test
cases generated follow a very specific formatting pattern. As it
happens, the latter turns out to be a safe assumption: I can't really
do anything about the former.

This has been integrated into the gen_test script