adds special handling of crust_abort
C doesn't have a notion of a bottom type, which means that we have to be
careful translating crust_abort (much like we have to be careful translating
return). We use a similar trick here: the call to crust_abort is transformed
into a block that "returns" an uninitialized variable of the correct type.
This is okay because execution never gets to the point where this uninitialized
value is ever used...

Given this I can probably finally remove the "clean_ir" pass from the cRep
portion of the compiler, but all things in time