changes to test gen
turns out that test scrubbing needs to happen in several passes: one to filter out
tests that fail due to attempting to access non-public functions, and another
for filtering out borrow test errors. Gen-test now runs the scrubber up to two
times and if translation fails on the third try the script terminates.

Also trying to pass in the failing lines as a huge comma-separated command line
argument wasn't a winning strategy: this commit fixes this