work around for assert
we can't directly translate the call to crust_assert into
assert() because this is secretly a macro that returns
a who-knows-what, not an integer (which is what we've typedef'd unit
to be). To fix this, we generate a call to a wrapper (in rust_intrinsics)
which does the assertion and then returns a sentinel unit value.