Stdlib minification, command line sanity
This commit significantly speeds up test generation by minimizing the
standard library before processing the generated tests. Preprocess
is not performant on very large files and proved to be a bottleneck.
I tracked this problem down to the sheer size of the file being parsed.

To address this, I used the fact that, when generating the public API,
we must walk all functions and types used by the module under test.
As a result, we know *exactly* which types, functions, statics,
and abstract functions that could possibly be used by the library
(and by extension, our test script). This pass in crust simply outputs
the names of these items.

A separate pass in Preprocess will now filter out the types and
functions now appearing in this list, producing a significantly
smaller IR representation of the module. As a result, test generation
time is MUCH MUCH faster.

As a part of this change, I also cleaned up how we manage command line
arguments, as the "several boolean flags" was becoming untenable. I also
changed how input is read into crust: stdin is the default UNLESS a filename
is provided on the command line. This appears to the standard for
most tools, although the special "-" argument is still accepted.