handles default trait implementations
Due to how we're currently handling default trait implementations,
we can get some ambiguity (both the overridden and default are found).
This resolves this ambiguity by looking for the magic "default" marker
on the function name and discarding any resolutions that are so marked
and returning the remaining function if it exists.

If it does not exist an exception is raised