scrubs defaults with no abstract fn
Resolving defaults requires abstract function definitions, but abstract
function definitions are only translated when the rustc plugin comes
across a call to an abstract function. This means that if a default method
is defined but never used Preprocess freaks out.

So this pass scrubs use_default commands that use bad abstract functions.
A more complete solution would be to add an abstract function if a default
method is discovered. I have code that does that, but one such abstract_fn
crashes the translator for abstract functions and I don't know how to
get around that.

The next commit will have that code in a comment, hopefully Stuart can
do something with it. But for the moment this gets the stdlib working again.