adds loop bound finding to test runner
It turns out that just doing "unwind=n" in cbmc isn't a great idea, because
some memory copy operations are modeled as loops, and what is a nice bound
for user loops (say 5) is WAY too small to copy a 32-byte structure.

So instead, the test runner asks cbmc to find all loops and bounds only those
that it knows are part of user code.

What constitutes user code?
Anything that isn't in mem::ptr, memcpy, or memmove.

This is conservative but extending this set is easy enough.