some prettifying of pp code
Adds a pre-processor pass that removes calls to drop_glue that it can
prove will not translate to anything interesting (i.e. primitives, refs,
and pointers).

This also means that several lifttemps are redundant. So another pp pass
removes these assignments.

This shouldn't have an effect on the actual code generated, except to make
it a little bit nicer to read (lipstick, pig, etc. etc.)