adds interesting-ness heuristic
This commit adds an interesting-ness heuristic.

A function in the public API is considered interesting IFF:
* it contains unsafe code that is not simply a call to abort
* it calls a private function that is itself interesting

This filters out functions with call graphs like:
pub:foo -> pub:bar (interesting)
as well as:
pub:foo -> priv:bar

But includes functions like:
* pub:foo (interesting)
* pub:foo -> priv:bar -> priv:baz (interesting)

A function is also included if it is not interesting but it produces
a type that is required to call an interesting function (think
iterators).

This heuristic yields an astonishing reduction in the number of test
cases generated. For instance, it cuts the number of ring_buf tests generated
from 90,000 to around 13,000