adds a fix-clone pass
it turns out that the auto-generated clone was taking a reference and just
blindly dereferencing it, despite this not at all being an okay thing to do
for zero-sized types (whose pointers can be absolute garbage).

So instead we patch up any clone methods to see if the type being cloned is
zero sized and return a sentinel un-init copy of that type.