fixes block-ret-fix pass
what irony! It turns out after desugaring the types of some blocks did not EXACTLY match,
even if the code was still type correct.

Unfortunately, the fix-block-pass required EXACT type equality to run the pass, so this code
moves the running on this pass higher up in the order of passes.